requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.12.3
pydantic-settings==2.11.0
email-validator==2.3.0
python-dotenv==1.1.1
